feat: created function to calculate worker vales
